fixed button
i have fixed the button to refresh the graph, and on the way to doing that i have fixed the tickbox generation.
the button was fixed by putting the graphGenerator function in a lambda and the tickboxes were fixed by making them a tkinter boolean instead of ints. The code works as intended now